src/main/java/JsonFlatMain.java
import java.io.BufferedReader;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.nio.charset.Charset;

public class JsonFlatMain {
    public static String readInput(InputStream in) {
        StringBuilder sb = new StringBuilder();
        try {
            BufferedReader reader = new BufferedReader(new InputStreamReader(in, Charset.defaultCharset()));
            String line;
            while ((line = reader.readLine()) != null) {
                sb.append(line);
            }
            reader.close();
        } catch (Exception e) {
            e.printStackTrace();
            return "";
        }
        return sb.toString();
    }

    public static String flatJson(String in) {
        return in;
    }

    public static void main(String[] args) {
        String input = readInput(System.in);
        String flatten = flatJson(input);
        System.out.println(flatten);
    }
}
